indexer-agent,indexer-common: Move Network module to indexer-common
- Move proofOfIndexing to StatusResolver
- Move transaction management to indexer-common
- Move allocationIdProof to @indexer-common/src/allocations/keys
