+++ packages/indexer-agent/src/indexer.ts
@@ -10,12 +10,10 @@ import {
   IndexerErrorCode,
   POIDisputeAttributes,
   IndexingStatusResolver,
-  BlockPointer,
   IndexingStatus,
   SubgraphIdentifierType,
   parseGraphQLIndexingStatus,
 } from '@graphprotocol/indexer-common'
-import pRetry from 'p-retry'
 
 const POI_DISPUTES_CONVERTERS_FROM_GRAPHQL: Record<
   keyof POIDisputeAttributes,
@@ -188,76 +186,6 @@ export class Indexer {
         { err },
       )
       throw err
-    }
-  }
-
-  async proofOfIndexing(
-    deployment: SubgraphDeploymentID,
-    block: BlockPointer,
-    indexerAddress: string,
-  ): Promise<string | undefined> {
-    try {
-      return await pRetry(
-        async attempt => {
-          const result = await this.statusResolver.statuses
-            .query(
-              gql`
-                query proofOfIndexing(
-                  $subgraph: String!
-                  $blockNumber: Int!
-                  $blockHash: String!
-                  $indexer: String!
-                ) {
-                  proofOfIndexing(
-                    subgraph: $subgraph
-                    blockNumber: $blockNumber
-                    blockHash: $blockHash
-                    indexer: $indexer
-                  )
-                }
-              `,
-              {
-                subgraph: deployment.ipfsHash,
-                blockNumber: block.number,
-                blockHash: block.hash,
-                indexer: indexerAddress,
-              },
-            )
-            .toPromise()
-
-          if (result.error) {
-            throw result.error
-          }
-          this.logger.trace('Reference PoI generated', {
-            indexer: this.indexerAddress,
-            subgraph: deployment.ipfsHash,
-            block: block,
-            proof: result.data.proofOfIndexing,
-            attempt,
-          })
-
-          return result.data.proofOfIndexing
-        },
-        {
-          retries: 10,
-          onFailedAttempt: err => {
-            this.logger.warn(`Proof of indexing could not be queried`, {
-              attempt: err.attemptNumber,
-              retriesLeft: err.retriesLeft,
-              err: err.message,
-            })
-          },
-        },
-      )
-    } catch (error) {
-      const err = indexerError(IndexerErrorCode.IE019, error)
-      this.logger.error(`Failed to query proof of indexing`, {
-        subgraph: deployment.ipfsHash,
-        blockHash: block,
-        indexer: this.indexerAddress,
-        err: err,
-      })
-      return undefined
     }
   }
