+++ packages/indexer-common/package.json
@@ -35,6 +35,7 @@
     "graphql-tag": "2.12.6",
     "morgan": "1.10.0",
     "ngeohash": "0.6.3",
+    "p-filter": "2.1.0",
     "p-retry": "4.6.1",
     "sequelize": "6.18.0",
     "ts-custom-error": "^3.2.0"
