+++ packages/indexer-common/src/indexer-management/models/index.ts
@@ -4,8 +4,8 @@ import { IndexingRuleModels, defineIndexingRuleModels } from './indexing-rule'
 import { CostModelModels, defineCostModelModels } from './cost-model'
 import { POIDisputeModels, definePOIDisputeModels } from './poi-dispute'
 
-export * from './indexing-rule'
 export * from './cost-model'
+export * from './indexing-rule'
 export * from './poi-dispute'
 
 export type IndexerManagementModels = IndexingRuleModels &
