+++ packages/indexer-common/src/network.ts
@@ -15,7 +15,6 @@ import {
   Epoch,
   INDEXER_ERROR_MESSAGES,
   indexerError,
-  IndexerError,
   IndexerErrorCode,
   INDEXING_RULE_GLOBAL,
   IndexingDecisionBasis,
@@ -26,69 +25,38 @@ import {
   Subgraph,
   SubgraphIdentifierType,
   SubgraphVersion,
+  TransactionManager,
   uniqueAllocationID,
 } from '@graphprotocol/indexer-common'
-import {
-  BigNumber,
-  BigNumberish,
-  ContractReceipt,
-  ContractTransaction,
-  providers,
-  Signer,
-  utils,
-  Wallet,
-} from 'ethers'
+import { BigNumber, providers, utils, Wallet } from 'ethers'
 import { strict as assert } from 'assert'
 import gql from 'graphql-tag'
 import geohash from 'ngeohash'
-import delay from 'delay'
 import pFilter from 'p-filter'
 import pRetry from 'p-retry'
+import { allocationIdProof } from './allocations/keys'
 
-const allocationIdProof = (
-  signer: Signer,
-  indexerAddress: string,
-  allocationId: string,
-): Promise<string> => {
-  const messageHash = utils.solidityKeccak256(
-    ['address', 'address'],
-    [indexerAddress, allocationId],
-  )
-  const messageHashBytes = utils.arrayify(messageHash)
-  return signer.signMessage(messageHashBytes)
-}
-
-enum TransactionType {
-  ZERO,
-  TWO,
-}
-interface TransactionConfig extends providers.TransactionRequest {
-  attempt: number
-  gasBump: BigNumber
-  type: TransactionType
-}
-
-export class Network {
-  networkSubgraph: NetworkSubgraph
-  contracts: NetworkContracts
-  indexerAddress: Address
-  indexerUrl: string
-  indexerGeoCoordinates: [string, string]
-  wallet: Wallet
-  logger: Logger
-  ethereum: providers.StaticJsonRpcProvider
-  paused: Eventual<boolean>
-  isOperator: Eventual<boolean>
+interface IndexerConfig {
+  url: string
+  geoCoordinates: [string, string]
   restakeRewards: boolean
   rebateClaimThreshold: BigNumber
   rebateClaimBatchThreshold: BigNumber
   rebateClaimMaxBatchSize: number
   poiDisputeMonitoring: boolean
   poiDisputableEpochs: number
-  gasIncreaseTimeout: number
-  gasIncreaseFactor: BigNumber
-  baseFeePerGasMax: number
-  maxTransactionAttempts: number
+}
+
+export class Network {
+  logger: Logger
+  networkSubgraph: NetworkSubgraph
+  contracts: NetworkContracts
+  indexerAddress: Address
+  ethereum: providers.StaticJsonRpcProvider
+  transactionManager: TransactionManager
+  indexerConfigs: IndexerConfig
+  indexerUrl: string
+  indexerGeoCoordinates: [string, string]
 
   private constructor(
     logger: Logger,
@@ -113,305 +81,33 @@ export class Network {
     maxTransactionAttempts: number,
   ) {
     this.logger = logger
-    this.wallet = wallet
     this.indexerAddress = indexerAddress
     this.indexerUrl = indexerUrl
     this.indexerGeoCoordinates = geoCoordinates
     this.contracts = contracts
     this.networkSubgraph = networkSubgraph
     this.ethereum = ethereum
-    this.paused = paused
-    this.isOperator = isOperator
-    this.restakeRewards = restakeRewards
-    this.rebateClaimThreshold = rebateClaimThreshold
-    this.rebateClaimBatchThreshold = rebateClaimBatchThreshold
-    this.rebateClaimMaxBatchSize = rebateClaimMaxBatchSize
-    this.poiDisputeMonitoring = poiDisputeMonitoring
-    this.poiDisputableEpochs = poiDisputableEpochs
-    this.gasIncreaseTimeout = gasIncreaseTimeout
-    this.gasIncreaseFactor = utils.parseUnits(gasIncreaseFactor.toString(), 3)
-    this.baseFeePerGasMax = baseFeePerGasMax
-    this.maxTransactionAttempts = maxTransactionAttempts
-  }
-
-  async executeTransaction(
-    gasEstimation: () => Promise<BigNumber>,
-    transaction: (gasLimit: BigNumberish) => Promise<ContractTransaction>,
-    logger: Logger,
-  ): Promise<ContractReceipt | 'paused' | 'unauthorized'> {
-    if (await this.paused.value()) {
-      logger.info(`Network is paused, skipping this action`)
-      return 'paused'
+    this.indexerConfigs = {
+      url: indexerUrl,
+      geoCoordinates: geoCoordinates,
+      restakeRewards: restakeRewards,
+      rebateClaimThreshold: rebateClaimThreshold,
+      rebateClaimBatchThreshold: rebateClaimBatchThreshold,
+      rebateClaimMaxBatchSize: rebateClaimMaxBatchSize,
+      poiDisputeMonitoring: poiDisputeMonitoring,
+      poiDisputableEpochs: poiDisputableEpochs,
     }
 
-    if (!(await this.isOperator.value())) {
-      logger.info(
-        `Not authorized as an operator for indexer, skipping this action`,
-      )
-      return 'unauthorized'
-    }
-
-    let pending = true
-    let output: providers.TransactionReceipt | undefined = undefined
-
-    const feeData = await this.waitForGasPricesBelowThreshold(logger)
-    const paddedGasLimit = Math.ceil((await gasEstimation()).toNumber() * 1.5)
-
-    const txPromise = transaction(paddedGasLimit)
-    let tx = await txPromise
-    let txRequest: providers.TransactionRequest | undefined = undefined
-
-    let txConfig: TransactionConfig = {
-      attempt: 1,
-      type: await this.transactionType(feeData),
-      gasBump: this.gasIncreaseFactor,
-      nonce: tx.nonce,
-      maxFeePerGas: tx.maxFeePerGas,
-      maxPriorityFeePerGas: tx.maxPriorityFeePerGas,
-      gasPrice: tx.gasPrice,
-      gasLimit: tx.gasLimit,
-    }
-
-    logger.info(`Sending transaction`, { txConfig })
-
-    while (pending) {
-      if (
-        this.maxTransactionAttempts !== 0 &&
-        txConfig.attempt > this.maxTransactionAttempts
-      ) {
-        logger.warn('Transaction retry limit reached, giving up', {
-          txConfig,
-        })
-        await delay(30000)
-        break
-      }
-
-      try {
-        if (txConfig.attempt > 1) {
-          logger.info('Resubmitting transaction', {
-            txConfig,
-          })
-          txRequest = {
-            value: tx.value,
-            to: tx.to,
-            data: tx.data,
-            chainId: tx.chainId,
-            from: tx.from,
-            nonce: txConfig.nonce,
-            gasPrice: txConfig.gasPrice,
-            maxPriorityFeePerGas: txConfig.maxPriorityFeePerGas,
-            maxFeePerGas: txConfig.maxFeePerGas,
-            gasLimit: txConfig.gasLimit,
-          }
-          tx = await this.wallet.sendTransaction(txRequest)
-        }
-
-        logger.info(`Transaction pending`, { tx: tx })
-
-        const receipt = await this.ethereum.waitForTransaction(
-          tx.hash,
-          3,
-          this.gasIncreaseTimeout,
-        )
-
-        if (receipt.status == 0) {
-          const revertReason = await this.getRevertReason(
-            logger,
-            txRequest as providers.TransactionRequest,
-          )
-          if (revertReason === 'out of gas') {
-            throw indexerError(IndexerErrorCode.IE050)
-          } else if (revertReason === 'unknown') {
-            throw indexerError(IndexerErrorCode.IE051)
-          } else {
-            throw indexerError(IndexerErrorCode.IE057)
-          }
-        }
-
-        logger.info(`Transaction successfully included in block`, {
-          tx: tx.hash,
-          receipt: receipt,
-        })
-        output = receipt
-        pending = false
-      } catch (error) {
-        txConfig = await this.updateTransactionConfig(logger, txConfig, error)
-        continue
-      }
-    }
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    return output!
-  }
-
-  async getRevertReason(
-    logger: Logger,
-    txRequest: providers.TransactionRequest,
-  ): Promise<string> {
-    let revertReason = 'unknown'
-    try {
-      const code = await this.ethereum.call(txRequest)
-      revertReason = utils.toUtf8String(`0x${code.substr(138)}`)
-    } catch (e) {
-      if (e.body.includes('out of gas')) {
-        revertReason = 'out of gas'
-      } else {
-        throw indexerError(IndexerErrorCode.IE051)
-      }
-    }
-    logger.warn('Transaction reverted:', { reason: revertReason })
-    return revertReason
-  }
-
-  async updateTransactionConfig(
-    logger: Logger,
-    txConfig: TransactionConfig,
-    error: Error | IndexerError,
-  ): Promise<TransactionConfig> {
-    logger.warning(
-      'Failed to send transaction, evaluating retry possibilities',
-      {
-        error: error.message,
-      },
+    this.transactionManager = new TransactionManager(
+      ethereum,
+      wallet,
+      paused,
+      isOperator,
+      gasIncreaseTimeout,
+      gasIncreaseFactor,
+      baseFeePerGasMax,
+      maxTransactionAttempts,
     )
-    if (error instanceof IndexerError) {
-      if (error.code == IndexerErrorCode.IE050) {
-        txConfig.gasLimit = BigNumber.from(txConfig.gasLimit)
-          .mul(txConfig.gasBump)
-          .div(1000)
-        txConfig.nonce = BigNumber.from(txConfig.nonce).add(1)
-      } else if (error.code == IndexerErrorCode.IE051) {
-        throw error
-      }
-    } else if (error instanceof Error) {
-      if (
-        error.message.includes(
-          'Transaction with the same hash was already imported',
-        ) ||
-        error.message.includes('nonce has already been used')
-      ) {
-        // This case typically indicates a successful transaction being retried.
-        // Let's introduce a 30 second delay to ensure the previous transaction has
-        // a chance to be mined and return to the reconciliation loop so the agent can reevaluate.
-        delay(30000)
-        throw indexerError(
-          IndexerErrorCode.IE058,
-          `Original transaction was not confirmed though it may have been successful`,
-        )
-      } else if (
-        error.message.includes(
-          'Transaction nonce is too low. Try incrementing the nonce.',
-        )
-      ) {
-        txConfig.nonce = BigNumber.from(txConfig.nonce).add(1)
-      } else if (
-        error.message.includes('Try increasing the fee') ||
-        error.message.includes('gas price supplied is too low') ||
-        error.message?.includes('timeout exceeded')
-      ) {
-        // Transaction timed out or failed due to a low gas price estimation, bump gas price and retry
-        if (txConfig.type === TransactionType.ZERO) {
-          txConfig.gasPrice = BigNumber.from(txConfig.gasPrice)
-            .mul(txConfig.gasBump)
-            .div(1000)
-        } else if (txConfig.type == TransactionType.TWO) {
-          // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-          txConfig.maxFeePerGas = BigNumber.from(txConfig.maxFeePerGas)
-            .mul(txConfig.gasBump)
-            .div(1000)
-          txConfig.maxPriorityFeePerGas = BigNumber.from(
-            txConfig.maxPriorityFeePerGas,
-          )
-            .mul(txConfig.gasBump)
-            .div(1000)
-        }
-      }
-    }
-    txConfig.attempt += 1
-    return txConfig
-  }
-  async transactionType(data: providers.FeeData): Promise<TransactionType> {
-    if (data.maxPriorityFeePerGas && data.maxFeePerGas) {
-      return TransactionType.TWO
-    } else if (data.gasPrice) {
-      return TransactionType.ZERO
-    } else {
-      throw new Error(
-        `Network fee data failed validation: gasPrice: ${data.gasPrice}, maxPriorityFeePerGas: ${data.maxPriorityFeePerGas}, maxFeePerGass: ${data.maxFeePerGas}`,
-      )
-    }
-  }
-  async waitForGasPricesBelowThreshold(
-    logger: Logger,
-  ): Promise<providers.FeeData> {
-    let attempt = 1
-    let aboveThreshold = true
-    let feeData = {
-      gasPrice: null,
-      maxFeePerGas: null,
-      maxPriorityFeePerGas: null,
-    } as providers.FeeData
-
-    while (aboveThreshold) {
-      feeData = await this.ethereum.getFeeData()
-      const type = await this.transactionType(feeData)
-      if (type === TransactionType.TWO) {
-        // Type 0x02 transaction
-        // This baseFeePerGas calculation is based off how maxFeePerGas is calculated in getFeeData()
-        // https://github.com/ethers-io/ethers.js/blob/68229ac0aff790b083717dc73cd84f38d32a3926/packages/abstract-provider/src.ts/index.ts#L247
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        const baseFeePerGas = feeData
-          .maxFeePerGas! // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-          .sub(feeData.maxPriorityFeePerGas!)
-          .div(2)
-        if (baseFeePerGas.toNumber() >= this.baseFeePerGasMax) {
-          if (attempt == 1) {
-            logger.warning(
-              `Max base fee per gas has been reached, waiting until the base fee falls below to resume transaction execution.`,
-              { maxBaseFeePerGas: this.baseFeePerGasMax, baseFeePerGas },
-            )
-          } else {
-            logger.info(
-              `Base gas fee per gas estimation still above max threshold`,
-              {
-                maxBaseFeePerGas: this.baseFeePerGasMax,
-                baseFeePerGas,
-                priceEstimateAttempt: attempt,
-              },
-            )
-          }
-          await delay(30000)
-          attempt++
-        } else {
-          aboveThreshold = false
-          feeData.gasPrice = null
-        }
-      } else if (type === TransactionType.ZERO) {
-        // Legacy transaction type
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        if (feeData.gasPrice!.toNumber() >= this.baseFeePerGasMax) {
-          if (attempt == 1) {
-            logger.warning(
-              `Max gas price has been reached, waiting until gas price estimates fall below to resume transaction execution.`,
-              {
-                baseFeePerGasMax: this.baseFeePerGasMax,
-                currentGasPriceEstimate: feeData.gasPrice,
-              },
-            )
-          } else {
-            logger.info(`Gas price estimation still above max threshold`, {
-              baseFeePerGasMax: this.baseFeePerGasMax,
-              currentGasPriceEstimate: feeData.gasPrice,
-              priceEstimateAttempt: attempt,
-            })
-          }
-          await delay(30000)
-          attempt++
-        } else {
-          aboveThreshold = false
-        }
-      }
-    }
-    return feeData
   }
 
   static async create(
@@ -440,17 +136,8 @@ export class Network {
       operator: wallet.address,
     })
 
-    const paused = await monitorNetworkPauses(
-      logger,
-      contracts,
-      networkSubgraph,
-    )
-    const isOperator = await monitorIsOperator(
-      logger,
-      contracts,
-      indexerAddress,
-      wallet,
-    )
+    const paused = await monitorNetworkPauses(logger, contracts, networkSubgraph)
+    const isOperator = await monitorIsOperator(logger, contracts, indexerAddress, wallet)
 
     return new Network(
       logger,
@@ -498,11 +185,7 @@ export class Network {
       try {
         const result = await this.networkSubgraph.query(
           gql`
-            query subgraphs(
-              $first: Int!
-              $lastId: String!
-              $subgraphs: [String!]!
-            ) {
+            query subgraphs($first: Int!, $lastId: String!, $subgraphs: [String!]!) {
               subgraphs(
                 where: { id_gt: $lastId, id_in: $subgraphs }
                 orderBy: id
@@ -541,9 +224,7 @@ export class Network {
               return {
                 version: versionItem.version,
                 createdAt: versionItem.createdAt,
-                deployment: new SubgraphDeploymentID(
-                  versionItem.subgraphDeployment.id,
-                ),
+                deployment: new SubgraphDeploymentID(versionItem.subgraphDeployment.id),
               } as SubgraphVersion
             },
           )
@@ -552,16 +233,11 @@ export class Network {
 
         // In the case of a fresh graph network there will be no published subgraphs, handle gracefully
         if (results.length == 0 && queryProgress.fetched == 0) {
-          this.logger.warn(
-            'No subgraph deployments found matching provided ids',
-            {
-              retriesRemaining: queryProgress.retriesRemaining,
-              subgraphIds: ids,
-            },
-          )
-          throw new Error(
-            `No subgraph deployments found matching provided ids: ${ids}`,
-          )
+          this.logger.warn('No subgraph deployments found matching provided ids', {
+            retriesRemaining: queryProgress.retriesRemaining,
+            subgraphIds: ids,
+          })
+          throw new Error(`No subgraph deployments found matching provided ids: ${ids}`)
         }
 
         queryProgress.exhausted = results.length < queryProgress.first
@@ -592,9 +268,7 @@ export class Network {
   async deploymentsWorthAllocatingTowards(
     rules: IndexingRuleAttributes[],
   ): Promise<SubgraphDeploymentID[]> {
-    const globalRule = rules.find(
-      rule => rule.identifier === INDEXING_RULE_GLOBAL,
-    )
+    const globalRule = rules.find((rule) => rule.identifier === INDEXING_RULE_GLOBAL)
 
     const deployments = []
     const queryProgress = {
@@ -604,9 +278,7 @@ export class Network {
       exhausted: false,
       retriesRemaining: 10,
     }
-    this.logger.info(
-      `Query subgraph deployments in batches of ${queryProgress.first}`,
-    )
+    this.logger.info(`Query subgraph deployments in batches of ${queryProgress.first}`)
 
     while (!queryProgress.exhausted) {
       this.logger.trace(`Query subgraph deployments`, {
@@ -663,11 +335,10 @@ export class Network {
               const deploymentRule =
                 rules
                   .filter(
-                    rule =>
-                      rule.identifierType == SubgraphIdentifierType.DEPLOYMENT,
+                    (rule) => rule.identifierType == SubgraphIdentifierType.DEPLOYMENT,
                   )
                   .find(
-                    rule =>
+                    (rule) =>
                       new SubgraphDeploymentID(rule.identifier).toString() ===
                       deployment.id,
                   ) || globalRule
@@ -684,15 +355,9 @@ export class Network {
 
               if (deploymentRule) {
                 const stakedTokens = BigNumber.from(deployment.stakedTokens)
-                const signalledTokens = BigNumber.from(
-                  deployment.signalledTokens,
-                )
-                const avgQueryFees = BigNumber.from(
-                  deployment.queryFeesAmount,
-                ).div(
-                  BigNumber.from(
-                    Math.max(1, deployment.indexerAllocations.length),
-                  ),
+                const signalledTokens = BigNumber.from(deployment.signalledTokens)
+                const avgQueryFees = BigNumber.from(deployment.queryFeesAmount).div(
+                  BigNumber.from(Math.max(1, deployment.indexerAllocations.length)),
                 )
 
                 this.logger.trace('Deciding whether to allocate and index', {
@@ -716,32 +381,23 @@ export class Network {
                       ? BigNumber.from(deploymentRule.maxSignal).toString()
                       : null,
                     minAverageQueryFees: deploymentRule.minAverageQueryFees
-                      ? BigNumber.from(
-                          deploymentRule.minAverageQueryFees,
-                        ).toString()
+                      ? BigNumber.from(deploymentRule.minAverageQueryFees).toString()
                       : null,
                     requireSupported: deploymentRule.requireSupported,
                   },
                 })
 
                 // Reject unsupported subgraph by default
-                if (
-                  deployment.deniedAt > 0 &&
-                  deploymentRule.requireSupported
-                ) {
+                if (deployment.deniedAt > 0 && deploymentRule.requireSupported) {
                   return false
                 }
 
                 // Skip the indexing rules checks if the decision basis is 'always', 'never', or 'offchain'
-                if (
-                  deploymentRule?.decisionBasis === IndexingDecisionBasis.ALWAYS
-                ) {
+                if (deploymentRule?.decisionBasis === IndexingDecisionBasis.ALWAYS) {
                   return true
                 } else if (
-                  deploymentRule?.decisionBasis ===
-                    IndexingDecisionBasis.NEVER ||
-                  deploymentRule?.decisionBasis ===
-                    IndexingDecisionBasis.OFFCHAIN
+                  deploymentRule?.decisionBasis === IndexingDecisionBasis.NEVER ||
+                  deploymentRule?.decisionBasis === IndexingDecisionBasis.OFFCHAIN
                 ) {
                   return false
                 }
@@ -774,12 +430,9 @@ export class Network {
         })
         if (queryProgress.retriesRemaining <= 0) {
           const error = indexerError(IndexerErrorCode.IE009, err.message)
-          this.logger.error(
-            `Failed to query subgraph deployments worth indexing`,
-            {
-              error,
-            },
-          )
+          this.logger.error(`Failed to query subgraph deployments worth indexing`, {
+            error,
+          })
           throw error
         }
       }
@@ -847,12 +500,9 @@ export class Network {
       return result.data.indexer.allocations.map(parseGraphQLAllocation)
     } catch (error) {
       const err = indexerError(IndexerErrorCode.IE010, error)
-      this.logger.error(
-        `Failed to query indexer's recently closed allocations`,
-        {
-          err,
-        },
-      )
+      this.logger.error(`Failed to query indexer's recently closed allocations`, {
+        err,
+      })
       throw err
     }
   }
@@ -928,10 +578,7 @@ export class Network {
       const result = await this.networkSubgraph.query(
         gql`
           query allocations($indexer: String!, $status: AllocationStatus!) {
-            allocations(
-              where: { indexer: $indexer, status: $status }
-              first: 1000
-            ) {
+            allocations(where: { indexer: $indexer, status: $status }, first: 1000) {
               id
               indexer {
                 id
@@ -1008,7 +655,7 @@ export class Network {
         {
           indexer: this.indexerAddress.toLocaleLowerCase(),
           disputableEpoch,
-          minimumQueryFeesCollected: this.rebateClaimThreshold.toString(),
+          minimumQueryFeesCollected: this.indexerConfigs.rebateClaimThreshold.toString(),
         },
       )
 
@@ -1023,25 +670,24 @@ export class Network {
         BigNumber.from(0),
       )
 
-      const parsedAllocs: Allocation[] = result.data.allocations.map(
-        parseGraphQLAllocation,
-      )
+      const parsedAllocs: Allocation[] =
+        result.data.allocations.map(parseGraphQLAllocation)
 
       // If the total fees claimable do not meet the minimum required for batching, return an empty array
       if (
         parsedAllocs.length > 0 &&
-        totalFees.lt(this.rebateClaimBatchThreshold)
+        totalFees.lt(this.indexerConfigs.rebateClaimBatchThreshold)
       ) {
         this.logger.info(
           `Allocation rebate batch value does not meet minimum for claiming`,
           {
             batchValueGRT: formatGRT(totalFees),
             rebateClaimBatchThreshold: formatGRT(
-              this.rebateClaimBatchThreshold,
+              this.indexerConfigs.rebateClaimBatchThreshold,
             ),
-            rebateClaimMaxBatchSize: this.rebateClaimMaxBatchSize,
+            rebateClaimMaxBatchSize: this.indexerConfigs.rebateClaimMaxBatchSize,
             batchSize: parsedAllocs.length,
-            allocations: parsedAllocs.map(allocation => {
+            allocations: parsedAllocs.map((allocation) => {
               return {
                 allocation: allocation.id,
                 deployment: allocation.subgraphDeployment.id.display,
@@ -1071,8 +717,8 @@ export class Network {
     minimumAllocation: number,
   ): Promise<Allocation[]> {
     const logger = this.logger.child({ component: 'PoI Monitor' })
-    if (!this.poiDisputeMonitoring) {
-      logger.debug('PoI monitoring disabled, skipping')
+    if (!this.indexerConfigs.poiDisputeMonitoring) {
+      logger.trace('PoI monitoring disabled, skipping')
       return Promise.resolve([])
     }
 
@@ -1085,7 +731,8 @@ export class Network {
 
     try {
       const zeroPOI = utils.hexlify(Array(32).fill(0))
-      const disputableEpoch = currentEpoch.toNumber() - this.poiDisputableEpochs
+      const disputableEpoch =
+        currentEpoch.toNumber() - this.indexerConfigs.poiDisputableEpochs
       let lastCreatedAt = 0
       while (dataRemaining) {
         const result = await this.networkSubgraph.query(
@@ -1129,7 +776,7 @@ export class Network {
             }
           `,
           {
-            deployments: deployments.map(subgraph => subgraph.bytes32),
+            deployments: deployments.map((subgraph) => subgraph.bytes32),
             minimumAllocation,
             disputableEpoch,
             createdAt: lastCreatedAt,
@@ -1144,44 +791,38 @@ export class Network {
           dataRemaining = false
         } else {
           lastCreatedAt = result.data.allocations.slice(-1)[0].createdAt
-          const parsedResult: Allocation[] = result.data.allocations.map(
-            parseGraphQLAllocation,
-          )
+          const parsedResult: Allocation[] =
+            result.data.allocations.map(parseGraphQLAllocation)
           allocations = allocations.concat(parsedResult)
         }
       }
 
       // Get the unique set of dispute epochs to reduce the work fetching epoch start block hashes in the next step
       let disputableEpochs = await this.epochs([
-        ...allocations.reduce(
-          (epochNumbers: Set<number>, allocation: Allocation) => {
-            epochNumbers.add(allocation.closedAtEpoch)
-            epochNumbers.add(allocation.closedAtEpoch - 1)
-            return epochNumbers
-          },
-          new Set(),
-        ),
+        ...allocations.reduce((epochNumbers: Set<number>, allocation: Allocation) => {
+          epochNumbers.add(allocation.closedAtEpoch)
+          epochNumbers.add(allocation.closedAtEpoch - 1)
+          return epochNumbers
+        }, new Set()),
       ])
 
       disputableEpochs = await Promise.all(
         disputableEpochs.map(async (epoch: Epoch): Promise<Epoch> => {
           // TODO: May need to retry or skip epochs where obtaining start block fails
-          epoch.startBlockHash = (
-            await this.ethereum.getBlock(epoch?.startBlock)
-          )?.hash
+          epoch.startBlockHash = (await this.ethereum.getBlock(epoch?.startBlock))?.hash
           return epoch
         }),
       )
 
       return await Promise.all(
-        allocations.map(async allocation => {
+        allocations.map(async (allocation) => {
           // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
           allocation.closedAtEpochStartBlockHash = disputableEpochs.find(
-            epoch => epoch.id == allocation.closedAtEpoch,
+            (epoch) => epoch.id == allocation.closedAtEpoch,
           )!.startBlockHash
           // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
           allocation.previousEpochStartBlockHash = disputableEpochs.find(
-            epoch => epoch.id == allocation.closedAtEpoch - 1,
+            (epoch) => epoch.id == allocation.closedAtEpoch - 1,
           )!.startBlockHash
           return allocation
         }),
@@ -1231,16 +872,84 @@ export class Network {
     }
   }
 
+  // async resolvePOI(
+  //   contracts: NetworkContracts,
+  //   transactionManager: TransactionManager,
+  //   indexingStatusResolver: IndexingStatusResolver,
+  //   allocation: Allocation,
+  //   poi: string | undefined,
+  //   force: boolean,
+  // ): Promise<string> {
+  //   // poi = undefined, force=true -- submit even if poi is 0x0
+  //   // poi = defined,   force=true ---> no generatedPOI needed, just submit the POI supplied (with some sanitation?)
+  //   // poi = undefined, force=false -- submit with generated POI if one available
+  //   // poi = defined,   force=false -- submit user defined POI only if generated POI matches
+  //   switch (force) {
+  //     case true:
+  //       switch (!!poi) {
+  //         case true:
+  //           // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  //           return poi!
+  //         case false:
+  //           return (
+  //             (await indexingStatusResolver.proofOfIndexing(
+  //               allocation.subgraphDeployment.id,
+  //               await transactionManager.ethereum.getBlock(
+  //                 (await contracts.epochManager.currentEpochBlock()).toNumber(),
+  //               ),
+  //               allocation.indexer,
+  //             )) || utils.hexlify(Array(32).fill(0))
+  //           )
+  //       }
+  //       break
+  //     case false: {
+  //       // Obtain the start block of the current epoch
+  //       const epochStartBlockNumber = await contracts.epochManager.currentEpochBlock()
+  //       const epochStartBlock = await transactionManager.ethereum.getBlock(
+  //         epochStartBlockNumber.toNumber(),
+  //       )
+  //       const generatedPOI = await indexingStatusResolver.proofOfIndexing(
+  //         allocation.subgraphDeployment.id,
+  //         epochStartBlock,
+  //         allocation.indexer,
+  //       )
+  //       switch (poi == generatedPOI) {
+  //         case true:
+  //           if (poi == undefined) {
+  //             const deploymentStatus = await indexingStatusResolver.indexingStatus([
+  //               allocation.subgraphDeployment.id,
+  //             ])
+  //             throw new Error(`POI not available for deployment at current epoch start block.
+  //             currentEpochStartBlock: ${epochStartBlockNumber}
+  //             deploymentStatus: ${deploymentStatus}`)
+  //           } else {
+  //             return poi
+  //           }
+  //         case false:
+  //           if (poi == undefined && generatedPOI !== undefined) {
+  //             return generatedPOI
+  //           } else if (poi !== undefined && generatedPOI == undefined) {
+  //             return poi
+  //           }
+  //           throw new Error(`User provided PoKI does not match reference fetched from the graph-node. Use '--force' to bypass this POI accuracy check.
+  //             POI: ${poi},
+  //             referencePOI: ${generatedPOI}`)
+  //       }
+  //     }
+  //   }
+  // }
+
+  // Start of SEND functions
   async register(): Promise<void> {
     const geoHash = geohash.encode(
-      +this.indexerGeoCoordinates[0],
-      +this.indexerGeoCoordinates[1],
+      +this.indexerConfigs.geoCoordinates[0],
+      +this.indexerConfigs.geoCoordinates[1],
     )
 
     const logger = this.logger.child({
       address: this.indexerAddress,
-      url: this.indexerUrl,
-      geoCoordinates: this.indexerGeoCoordinates,
+      url: this.indexerConfigs.url,
+      geoCoordinates: this.indexerConfigs.geoCoordinates,
       geoHash,
     })
 
@@ -1251,41 +960,33 @@ export class Network {
 
           // Register the indexer (only if it hasn't been registered yet or
           // if its URL is different from what is registered on chain)
-          const isRegistered =
-            await this.contracts.serviceRegistry.isRegistered(
-              this.indexerAddress,
-            )
+          const isRegistered = await this.contracts.serviceRegistry.isRegistered(
+            this.indexerAddress,
+          )
           if (isRegistered) {
             const service = await this.contracts.serviceRegistry.services(
               this.indexerAddress,
             )
-            if (
-              service.url === this.indexerUrl &&
-              service.geohash === geoHash
-            ) {
-              if (await this.isOperator.value()) {
-                logger.info(
-                  `Indexer already registered, operator status already granted`,
-                )
+            if (service.url === this.indexerConfigs.url && service.geohash === geoHash) {
+              if (await this.transactionManager.isOperator.value()) {
+                logger.info(`Indexer already registered, operator status already granted`)
                 return
               } else {
-                logger.info(
-                  `Indexer already registered, operator status not yet granted`,
-                )
+                logger.info(`Indexer already registered, operator status not yet granted`)
               }
             }
           }
-          const receipt = await this.executeTransaction(
+          const receipt = await this.transactionManager.executeTransaction(
             () =>
               this.contracts.serviceRegistry.estimateGas.registerFor(
                 this.indexerAddress,
-                this.indexerUrl,
+                this.indexerConfigs.url,
                 geoHash,
               ),
-            gasLimit =>
+            (gasLimit) =>
               this.contracts.serviceRegistry.registerFor(
                 this.indexerAddress,
-                this.indexerUrl,
+                this.indexerConfigs.url,
                 geoHash,
                 {
                   gasLimit,
@@ -1297,11 +998,9 @@ export class Network {
             return
           }
           const events = receipt.events || receipt.logs
-          const event = events.find(event =>
+          const event = events.find((event) =>
             event.topics.includes(
-              this.contracts.serviceRegistry.interface.getEventTopic(
-                'ServiceRegistered',
-              ),
+              this.contracts.serviceRegistry.interface.getEventTopic('ServiceRegistered'),
             ),
           )
           assert.ok(event)
@@ -1327,12 +1026,9 @@ export class Network {
     const logger = this.logger.child({ deployment: deployment.display })
 
     if (amount.lt('0')) {
-      logger.warn(
-        'Cannot allocate a negative amount of GRT, skipping this allocation',
-        {
-          amount: amount.toString(),
-        },
-      )
+      logger.warn('Cannot allocate a negative amount of GRT, skipping this allocation', {
+        amount: amount.toString(),
+      })
       return
     }
 
@@ -1374,10 +1070,10 @@ export class Network {
 
       // Obtain a unique allocation ID
       const { allocationSigner, allocationId } = uniqueAllocationID(
-        this.wallet.mnemonic.phrase,
+        this.transactionManager.wallet.mnemonic.phrase,
         currentEpoch.toNumber(),
         deployment,
-        activeAllocations.map(allocation => allocation.id),
+        activeAllocations.map((allocation) => allocation.id),
       )
 
       // Double-check whether the allocationID already exists on chain, to
@@ -1387,9 +1083,7 @@ export class Network {
       //     enum AllocationState { Null, Active, Closed, Finalized, Claimed }
       //
       // in the contracts.
-      const state = await this.contracts.staking.getAllocationState(
-        allocationId,
-      )
+      const state = await this.contracts.staking.getAllocationState(allocationId)
       if (state !== 0) {
         logger.debug(`Skipping allocation as it already exists onchain`, {
           indexer: this.indexerAddress,
@@ -1405,7 +1099,7 @@ export class Network {
         allocation: allocationId,
       })
 
-      const receipt = await this.executeTransaction(
+      const receipt = await this.transactionManager.executeTransaction(
         async () =>
           this.contracts.staking.estimateGas.allocateFrom(
             this.indexerAddress,
@@ -1413,24 +1107,16 @@ export class Network {
             amount,
             allocationId,
             utils.hexlify(Array(32).fill(0)),
-            await allocationIdProof(
-              allocationSigner,
-              this.indexerAddress,
-              allocationId,
-            ),
+            await allocationIdProof(allocationSigner, this.indexerAddress, allocationId),
           ),
-        async gasLimit =>
+        async (gasLimit) =>
           this.contracts.staking.allocateFrom(
             this.indexerAddress,
             deployment.bytes32,
             amount,
             allocationId,
             utils.hexlify(Array(32).fill(0)),
-            await allocationIdProof(
-              allocationSigner,
-              this.indexerAddress,
-              allocationId,
-            ),
+            await allocationIdProof(allocationSigner, this.indexerAddress, allocationId),
             { gasLimit },
           ),
         logger.child({ action: 'allocate' }),
@@ -1508,21 +1194,15 @@ export class Network {
       //     enum AllocationState { Null, Active, Closed, Finalized, Claimed }
       //
       // in the contracts.
-      const state = await this.contracts.staking.getAllocationState(
-        allocation.id,
-      )
+      const state = await this.contracts.staking.getAllocationState(allocation.id)
       if (state !== 1) {
         logger.info(`Allocation has already been closed`)
         return true
       }
 
-      const receipt = await this.executeTransaction(
-        () =>
-          this.contracts.staking.estimateGas.closeAllocation(
-            allocation.id,
-            poi,
-          ),
-        gasLimit =>
+      const receipt = await this.transactionManager.executeTransaction(
+        () => this.contracts.staking.estimateGas.closeAllocation(allocation.id, poi),
+        (gasLimit) =>
           this.contracts.staking.closeAllocation(allocation.id, poi, {
             gasLimit,
           }),
@@ -1549,7 +1229,7 @@ export class Network {
       logger.info(
         `${allocations.length} allocations are eligible for rebate pool claims`,
         {
-          allocations: allocations.map(allocation => {
+          allocations: allocations.map((allocation) => {
             return {
               allocation: allocation.id,
               deployment: allocation.subgraphDeployment.id.display,
@@ -1558,53 +1238,46 @@ export class Network {
               createdAtBlockHash: allocation.createdAtBlockHash,
             }
           }),
-          restakeRewards: this.restakeRewards,
+          restakeRewards: this.indexerConfigs.restakeRewards,
         },
       )
 
       // Filter out already-claimed and still-active allocations
-      allocations = await pFilter(
-        allocations,
-        async (allocation: Allocation) => {
-          // Double-check whether the allocation is claimed to
-          // avoid unnecessary transactions.
-          // Note: We're checking the allocation state here, which is defined as
-          //
-          //     enum AllocationState { Null, Active, Closed, Finalized, Claimed }
-          //
-          // in the contracts.
-          const state = await this.contracts.staking.getAllocationState(
-            allocation.id,
+      allocations = await pFilter(allocations, async (allocation: Allocation) => {
+        // Double-check whether the allocation is claimed to
+        // avoid unnecessary transactions.
+        // Note: We're checking the allocation state here, which is defined as
+        //
+        //     enum AllocationState { Null, Active, Closed, Finalized, Claimed }
+        //
+        // in the contracts.
+        const state = await this.contracts.staking.getAllocationState(allocation.id)
+        if (state === 4) {
+          logger.trace(
+            `Allocation rebate rewards already claimed, ignoring ${allocation.id}.`,
           )
-          if (state === 4) {
-            logger.trace(
-              `Allocation rebate rewards already claimed, ignoring ${allocation.id}.`,
-            )
-            return false
-          }
-          if (state === 1) {
-            logger.trace(`Allocation still active, ignoring ${allocation.id}.`)
-            return false
-          }
-          return true
-        },
-      )
+          return false
+        }
+        if (state === 1) {
+          logger.trace(`Allocation still active, ignoring ${allocation.id}.`)
+          return false
+        }
+        return true
+      })
 
       // Max claims per batch should roughly be equal to average gas per claim / block gas limit
       // On-chain data shows an average of 120k gas per claim and the block gas limit is 15M
       // We get at least 21k gas savings per inclusion of a claim in a batch
       // A reasonable upper bound for this value is 200 assuming the system has the memory
       // requirements to construct the transaction
-      const maxClaimsPerBatch = this.rebateClaimMaxBatchSize
+      const maxClaimsPerBatch = this.indexerConfigs.rebateClaimMaxBatchSize
 
       // When we construct the batch, we sort desc by query fees collected
       // in order to maximise the value of the truncated batch
       // more query fees collected should mean higher value rebates
       const allocationIds = allocations
-        .sort((x, y) =>
-          y.queryFeesCollected?.gt(x.queryFeesCollected || 0) ? 1 : -1,
-        )
-        .map(allocation => allocation.id)
+        .sort((x, y) => (y.queryFeesCollected?.gt(x.queryFeesCollected || 0) ? 1 : -1))
+        .map((allocation) => allocation.id)
         .slice(0, maxClaimsPerBatch)
 
       if (allocationIds.length === 0) {
@@ -1618,16 +1291,20 @@ export class Network {
       }
 
       // Claim the earned value from the rebate pool, returning it to the indexers stake
-      const receipt = await this.executeTransaction(
+      const receipt = await this.transactionManager.executeTransaction(
         () =>
           this.contracts.staking.estimateGas.claimMany(
             allocationIds,
-            this.restakeRewards,
+            this.indexerConfigs.restakeRewards,
           ),
-        gasLimit =>
-          this.contracts.staking.claimMany(allocationIds, this.restakeRewards, {
-            gasLimit,
-          }),
+        (gasLimit) =>
+          this.contracts.staking.claimMany(
+            allocationIds,
+            this.indexerConfigs.restakeRewards,
+            {
+              gasLimit,
+            },
+          ),
         logger.child({ action: 'claimMany' }),
       )
       if (receipt === 'paused' || receipt === 'unauthorized') {
@@ -1709,9 +1386,7 @@ export class Network {
 
       // When reallocating, we will first close the old allocation and free up the GRT in that allocation
       // This GRT will be available in addition to freeStake for the new allocation
-      const postCloseFreeStake = freeStake.add(
-        existingAllocation.allocatedTokens,
-      )
+      const postCloseFreeStake = freeStake.add(existingAllocation.allocatedTokens)
 
       // If there isn't enough left for allocating, abort
       if (postCloseFreeStake.lt(amount)) {
@@ -1732,13 +1407,12 @@ export class Network {
       logger.debug('Obtain a unique Allocation ID')
 
       // Obtain a unique allocation ID
-      const { allocationSigner, allocationId: newAllocationId } =
-        uniqueAllocationID(
-          this.wallet.mnemonic.phrase,
-          currentEpoch.toNumber(),
-          deployment,
-          activeAllocations.map(allocation => allocation.id),
-        )
+      const { allocationSigner, allocationId: newAllocationId } = uniqueAllocationID(
+        this.transactionManager.wallet.mnemonic.phrase,
+        currentEpoch.toNumber(),
+        deployment,
+        activeAllocations.map((allocation) => allocation.id),
+      )
 
       // Double-check whether the allocationID already exists on chain, to
       // avoid unnecessary transactions.
@@ -1747,8 +1421,9 @@ export class Network {
       //     enum AllocationState { Null, Active, Closed, Finalized, Claimed }
       //
       // in the contracts.
-      const newAllocationState =
-        await this.contracts.staking.getAllocationState(newAllocationId)
+      const newAllocationState = await this.contracts.staking.getAllocationState(
+        newAllocationId,
+      )
       if (newAllocationState !== 0) {
         logger.warn(`Skipping Allocation as it already exists onchain`, {
           indexer: this.indexerAddress,
@@ -1774,7 +1449,7 @@ export class Network {
         proof,
       })
 
-      const receipt = await this.executeTransaction(
+      const receipt = await this.transactionManager.executeTransaction(
         async () =>
           this.contracts.staking.estimateGas.closeAndAllocate(
             existingAllocation.id,
@@ -1786,7 +1461,7 @@ export class Network {
             utils.hexlify(Array(32).fill(0)), // metadata
             proof,
           ),
-        async gasLimit =>
+        async (gasLimit) =>
           this.contracts.staking.closeAndAllocate(
             existingAllocation.id,
             poi,
@@ -1863,13 +1538,13 @@ export class Network {
       createdAtEpoch: allocation.createdAtEpoch,
       closedAtEpoch: allocation.closedAtEpoch,
       createdAtBlockHash: allocation.createdAtBlockHash,
-      restakeRewards: this.restakeRewards,
+      restakeRewards: this.indexerConfigs.restakeRewards,
     })
     try {
       logger.info(`Claim tokens from the rebate pool for allocation`, {
         deployment: allocation.subgraphDeployment.id.display,
         allocation: allocation.id,
-        claimAmount: this.restakeRewards,
+        claimAmount: this.indexerConfigs.restakeRewards,
       })
 
       // Double-check whether the allocation is claimed to
@@ -1879,9 +1554,7 @@ export class Network {
       //     enum AllocationState { Null, Active, Closed, Finalized, Claimed }
       //
       // in the contracts.
-      const state = await this.contracts.staking.getAllocationState(
-        allocation.id,
-      )
+      const state = await this.contracts.staking.getAllocationState(allocation.id)
       if (state === 4) {
         logger.trace(`Allocation rebate rewards already claimed`)
         return true
@@ -1892,16 +1565,20 @@ export class Network {
       }
 
       // Claim the earned value from the rebate pool, returning it to the indexers stake
-      const receipt = await this.executeTransaction(
+      const receipt = await this.transactionManager.executeTransaction(
         () =>
           this.contracts.staking.estimateGas.claim(
             allocation.id,
-            this.restakeRewards,
+            this.indexerConfigs.restakeRewards,
           ),
-        gasLimit =>
-          this.contracts.staking.claim(allocation.id, this.restakeRewards, {
-            gasLimit,
-          }),
+        (gasLimit) =>
+          this.contracts.staking.claim(
+            allocation.id,
+            this.indexerConfigs.restakeRewards,
+            {
+              gasLimit,
+            },
+          ),
         logger.child({ action: 'claim' }),
       )
       if (receipt === 'paused' || receipt === 'unauthorized') {
@@ -1926,7 +1603,7 @@ async function monitorNetworkPauses(
   networkSubgraph: NetworkSubgraph,
 ): Promise<Eventual<boolean>> {
   return timer(60_000)
-    .reduce(async currentlyPaused => {
+    .reduce(async (currentlyPaused) => {
       try {
         const result = await networkSubgraph.query(
           gql`
@@ -1948,17 +1625,14 @@ async function monitorNetworkPauses(
 
         return result.data.graphNetworks[0].isPaused
       } catch (err) {
-        logger.warn(
-          `Failed to check for network pause, assuming it has not changed`,
-          {
-            err: indexerError(IndexerErrorCode.IE007, err),
-            paused: currentlyPaused,
-          },
-        )
+        logger.warn(`Failed to check for network pause, assuming it has not changed`, {
+          err: indexerError(IndexerErrorCode.IE007, err),
+          paused: currentlyPaused,
+        })
         return currentlyPaused
       }
     }, await contracts.controller.paused())
-    .map(paused => {
+    .map((paused) => {
       logger.info(paused ? `Network paused` : `Network active`)
       return paused
     })
@@ -1978,12 +1652,9 @@ async function monitorIsOperator(
   }
 
   return timer(60_000)
-    .reduce(async isOperator => {
+    .reduce(async (isOperator) => {
       try {
-        return await contracts.staking.isOperator(
-          wallet.address,
-          indexerAddress,
-        )
+        return await contracts.staking.isOperator(wallet.address, indexerAddress)
       } catch (err) {
         logger.warn(
           `Failed to check operator status for indexer, assuming it has not changed`,
@@ -1992,7 +1663,7 @@ async function monitorIsOperator(
         return isOperator
       }
     }, await contracts.staking.isOperator(wallet.address, indexerAddress))
-    .map(isOperator => {
+    .map((isOperator) => {
       logger.info(
         isOperator
           ? `Have operator status for indexer`
